Type all repo ports with domain models, add organization_id to base Entity
- All 8 repo ports now accept domain models instead of dict:
  DepartmentRepoPort, VendorRepoPort, ProductRepoPort,
  WithdrawalRepoPort, MaterialRequestRepoPort, InvoiceRepoPort,
  InvoicingGateway, StockRepoPort
- Repo implementations accept Union[Model, dict] for backward compat
  with seed scripts, calling .model_dump() internally
- Added organization_id to kernel Entity base class so every entity
  carries multitenancy context natively
- Callers now pass domain models directly instead of dumping to dict
  and manually attaching organization_id
- Removed dead csv_import_service import and route from products API
- 78 tests passing, all API endpoints verified against seeded data

Made-with: Cursor

+++ backend/catalog/application/queries.py
@@ -87,19 +87,19 @@ async def update_product(product_id: str, updates: dict, conn=None) -> Optional[
     return await _prod_repo.update(product_id, updates, conn=conn)
 
 
-async def atomic_decrement_product(product_id: str, quantity: int, updated_at: str, conn=None) -> Optional[dict]:
+async def atomic_decrement_product(product_id: str, quantity: float, updated_at: str, conn=None) -> Optional[dict]:
     return await _prod_repo.atomic_decrement(product_id, quantity, updated_at, conn=conn)
 
 
-async def increment_product_quantity(product_id: str, quantity: int, updated_at: str, conn=None) -> None:
+async def increment_product_quantity(product_id: str, quantity: float, updated_at: str, conn=None) -> None:
     return await _prod_repo.increment_quantity(product_id, quantity, updated_at, conn=conn)
 
 
-async def add_product_quantity(product_id: str, quantity: int, updated_at: str, conn=None) -> Optional[dict]:
+async def add_product_quantity(product_id: str, quantity: float, updated_at: str, conn=None) -> Optional[dict]:
     return await _prod_repo.add_quantity(product_id, quantity, updated_at, conn=conn)
 
 
-async def atomic_adjust_product(product_id: str, quantity_delta: int, updated_at: str) -> Optional[dict]:
+async def atomic_adjust_product(product_id: str, quantity_delta: float, updated_at: str) -> Optional[dict]:
     return await _prod_repo.atomic_adjust(product_id, quantity_delta, updated_at)
